pyproject.toml
[project]
name = "llmevaluation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepeval>=3.9.7",
    "ollama>=0.6.1",
    "pandas>=3.0.2",
    "pytest>=9.0.3",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "rich>=14.3.4",
]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore:There is no current event loop:DeprecationWarning"
]
